requirements.txt
streamlit
crewai
crewai[tools]
langchain
python-dotenv
ipykernel
langchain_community
pypdf
langchain_text-splitters
langchain-openai
langchain-ollama
langchain-groq
litellm